Reset V3D_RENDER_SHADOW flag on file loading.
This flag shouldn't be saved in .blend files but because of strange reason
some files contains this flags which leads to object disappearing when
disabling buffer shadow in material. Current trunk shouldn't be saving
this flag and most probably buggy files were saved in some intermediate
version of blender where this flag can be be saved.

This should fix remained issue in #29255: Object invisible and weird polygons appearance

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@42225 954f8c5b-7b00-dc11-b283-0030488c597c