Previous fix for stabilization autoscale actually broke stabilizaiton itself (median point was sliding a bit from it's original position)
This happens because of how transformation for stabilization calculates:
image is scaling around it's center, so image translation should be recalculated
after scale was changed, but scale also depends on translation. That's where
tricky things happens. It's still not ideal for case of rotation, but before
fixing this issue better to figure out usecase and see if it's indeed
so needed to scale around image center (it might only be helpful to use
stabilization parameters in compositor nodes).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@44766 954f8c5b-7b00-dc11-b283-0030488c597c