Fix bug #31529 Remesh remove Apply as Shape button
Change modifier_sameTopology() to not not treat Nonconstructive
modifiers as having same topology. This function is only used to test
whether Apply as Shape is allowed.

The exact nature of "nonconstructive" modifiers is not documented, but
current list is remesh, fluidsim, decimate, mask, build, and
boolean. All of these modify topology, and should not be making making
shape keys.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@46805 954f8c5b-7b00-dc11-b283-0030488c597c