* Enabled OpenCL for the compositor
Tested on:
   - Windows32 ATI V5800
   - Linux64 GTX360M
   - Linux64 Quadro FX360M

The ATI is much more stable as the Nvidia platforms. We have tested
a different scheduling that will speed up for AMD (not checked in yet)

It compiles on all platforms, but fails on our MAC book Pro.
Black lines are produced on the top of a opencl workgroup.

By using localworksize we were able to remove these lines, but 
are not satisfied at this solution (so will not check this in yet).

Please everyone check if it works on your configuration and add 
Bugs when needed.

To test you need to add a BokehBlur to your scene and enable the
OpenCL flag in the property panel.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@48060 954f8c5b-7b00-dc11-b283-0030488c597c