Fix for metaballs used as dupli-object for particle
It used to be a dependency cycle which lead to incorrect or
missed tesselation on some circumstances.

Seems to be introduced in rev41627.

This commit seems to behaving properly on simple cases,
probably could fail in some other cases, so need to be
checked further.

Discovered when was looking into:
    #32034: Metaball used as render object(group) for particle will display wire only.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@48858 954f8c5b-7b00-dc11-b283-0030488c597c