Sequencer: avoid allocating image buffer on every modifier
This means that modifier would operate on buffer which was passed to it,
without creating copy of image buffer and operating on it.

All current modifiers fit into this model and if it would need to have
original buffer on modifier calculation, that particular modifier can
create copy.

Gives some percentage of boost.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@50392 954f8c5b-7b00-dc11-b283-0030488c597c